fix: refactor radspec overrides to its own module
